fix(cli): add missing preprocessing step for custom gene maps in CLI
Attempts to resolve #252
